search updated

+++ frontend/src/components/medical-appointment-history/medical-apointment-history.jsx
@@ -17,6 +17,7 @@ import {
 } from "./medical-apointment-history-styles";
 import { Input } from "../UI/input.styles";
 import moment from "moment";
+import Swal from "sweetalert2";
 
 const EditableRow = ({ label, value, isEditing, onChange, fieldName }) => (
   <TableRow>
@@ -31,11 +32,36 @@ const EditableRow = ({ label, value, isEditing, onChange, fieldName }) => (
   </TableRow>
 );
 
+const ROWS_CONFIG = {
+  citaInfo: [
+    { label: "Especialista", field: "especialista" },
+    { label: "Fecha", field: "fecha" },
+    { label: "Hora", field: "hora" },
+    { label: "Tipo de Servicio", field: "tipo_servicio" },
+    { label: "Estado", field: "estado" },
+  ],
+  pacienteInfo: [
+    { label: "Paciente Nombre", field: "paciente_nombre" },
+    { label: "Edad", field: "paciente_edad" },
+    { label: "Nacionalidad", field: "paciente_nacionalidad" },
+    { label: "Cédula", field: "paciente_cedula" },
+    { label: "Género", field: "paciente_genero" },
+  ],
+  infoAdicional: [
+    { label: "Dirección", field: "paciente_direccion" },
+    { label: "Tipo de Paciente", field: "tipo_paciente" },
+    { label: "Solicitante Nombre", field: "solicitante_nombre" },
+    { label: "Solicitante Apellido", field: "solicitante_apellido" },
+    { label: "Afiliación ARS", field: "afiliacion_ars" },
+    { label: "Centro Nombre", field: "centro_nombre" },
+  ],
+};
+
 const PreviousAppointments = () => {
   const [appointments, setAppointments] = useState([]);
   const [filteredAppointments, setFilteredAppointments] = useState([]);
   const [search, setSearch] = useState("");
-  const [openAppointmentId, setOpenAppointmentId] = useState(null); // Estado que indica la cita abierta
+  const [openAppointmentId, setOpenAppointmentId] = useState(null);
   const [editingAppointmentId, setEditingAppointmentId] = useState(null);
   const [editedAppointment, setEditedAppointment] = useState({});
 
@@ -49,24 +75,23 @@ const PreviousAppointments = () => {
       .catch((error) => console.error("Error fetching appointments:", error));
   }, []);
 
-  const handleSearch = (e) => {
-    const query = e.target.value.toLowerCase();
-    setSearch(query);
-
+  const handleSearch = () => {
+    const query = search.toLowerCase();
     setFilteredAppointments(
       query
-        ? appointments.filter((app) => {
-            return (
-              (app.centro_nombre &&
-                app.centro_nombre.toLowerCase().includes(query)) ||
-              (app.especialista &&
-                app.especialista.toLowerCase().includes(query)) ||
-              (app.paciente_nombre &&
-                app.paciente_nombre.toLowerCase().includes(query))
-            );
-          })
+        ? appointments.filter((app) =>
+            ["centro_nombre", "especialista", "paciente_nombre"].some(
+              (key) => app[key]?.toLowerCase().includes(query)
+            )
+          )
         : appointments
     );
+  };
+
+  const handleKeyPress = (e) => {
+    if (e.key === "Enter") {
+      handleSearch();
+    }
   };
 
   const toggleAppointmentDetails = (appointmentId) => {
@@ -117,20 +142,45 @@ const PreviousAppointments = () => {
   };
 
   const handleDelete = (appointmentId) => {
-    fetch("http://localhost:5000/api/cita/eliminar", {
-      method: "DELETE",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ _id: appointmentId }),
-    })
-      .then(() => {
-        setAppointments(
-          appointments.filter((app) => app._id !== appointmentId)
-        );
-        setFilteredAppointments(
-          filteredAppointments.filter((app) => app._id !== appointmentId)
-        );
-      })
-      .catch((error) => console.error("Error al eliminar la cita:", error));
+    Swal.fire({
+      title: "¿Estás seguro?",
+      text: "No podrás revertir esta acción",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#3085d6",
+      confirmButtonText: "Sí, eliminar",
+      cancelButtonText: "Cancelar",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch("http://localhost:5000/api/cita/eliminar", {
+          method: "DELETE",
+          headers: { "Content-Type": "application/json" },
+          body: JSON.stringify({ _id: appointmentId }),
+        })
+          .then(() => {
+            Swal.fire(
+              "Eliminada",
+              "La cita ha sido eliminada correctamente.",
+              "success"
+            );
+            setAppointments((prevAppointments) =>
+              prevAppointments.filter((app) => app._id !== appointmentId)
+            );
+            setFilteredAppointments((prevAppointments) =>
+              prevAppointments.filter((app) => app._id !== appointmentId)
+            );
+          })
+          .catch((error) => {
+            console.error("Error al eliminar la cita:", error);
+            Swal.fire(
+              "Error",
+              "Hubo un problema al eliminar la cita. Intenta de nuevo.",
+              "error"
+            );
+          });
+      }
+    });
   };
 
   return (
@@ -139,11 +189,16 @@ const PreviousAppointments = () => {
         <CardTitle>Citas Registradas</CardTitle>
       </CardHeader>
       <CardContent>
-        <Input
-          placeholder="Buscar por centro, doctor o paciente..."
-          value={search}
-          onChange={handleSearch}
-        />
+      <div style={{ marginBottom: "10px" }}>
+          <Input
+            placeholder="Buscar por centro, doctor o paciente..."
+            value={search}
+            onChange={(e) => setSearch(e.target.value)} 
+            onKeyDown={handleKeyPress} 
+            
+          />
+          <Button onClick={handleSearch}>Buscar</Button> 
+        </div>
         <div>
           {filteredAppointments.map((appointment) => (
             <div key={appointment._id}>
@@ -159,78 +214,31 @@ const PreviousAppointments = () => {
 
               {openAppointmentId === appointment._id && (
                 <TableContainer>
-                  <TableColumn>
-                    <TableTitle>Info Cita</TableTitle>
-                    {[
-                      { label: "Especialista", field: "especialista" },
-                      { label: "Fecha", field: "fecha" },
-                      { label: "Hora", field: "hora" },
-                      { label: "Tipo de Servicio", field: "tipo_servicio" },
-                      { label: "Estado", field: "estado" },
-                    ].map(({ label, field }) => (
-                      <EditableRow
-                        key={field}
-                        label={label}
-                        value={
-                          field === "fecha"
-                            ? moment(appointment[field]).format("DD/MM/YYYY")
-                            : appointment[field]
-                        } // Formateo solo la fecha
-                        isEditing={editingAppointmentId === appointment._id}
-                        onChange={handleInputChange}
-                        fieldName={field}
-                      />
-                    ))}
-                  </TableColumn>
-
-                  <TableColumn>
-                    <TableTitle>Info Paciente</TableTitle>
-                    {[
-                      { label: "Paciente Nombre", field: "paciente_nombre" },
-                      { label: "Edad", field: "paciente_edad" },
-                      { label: "Nacionalidad", field: "paciente_nacionalidad" },
-                      { label: "Cédula", field: "paciente_cedula" },
-                      { label: "Género", field: "paciente_genero" },
-                    ].map(({ label, field }) => (
-                      <EditableRow
-                        key={field}
-                        label={label}
-                        value={appointment[field]}
-                        isEditing={editingAppointmentId === appointment._id}
-                        onChange={handleInputChange}
-                        fieldName={field}
-                      />
-                    ))}
-                  </TableColumn>
-
-                  <TableColumn>
-                    <TableTitle>Info Adicional</TableTitle>
-                    {[
-                      { label: "Dirección", field: "paciente_direccion" },
-                      { label: "Tipo de Paciente", field: "tipo_paciente" },
-                      {
-                        label: "Solicitante Nombre",
-                        field: "solicitante_nombre",
-                      },
-                      {
-                        label: "Solicitante Apellido",
-                        field: "solicitante_apellido",
-                      },
-                      { label: "Afiliación ARS", field: "afiliacion_ars" },
-                      { label: "Centro Nombre", field: "centro_nombre" },
-                    ].map(({ label, field }) => (
-                      <EditableRow
-                        key={field}
-                        label={label}
-                        value={appointment[field]}
-                        isEditing={editingAppointmentId === appointment._id}
-                        onChange={handleInputChange}
-                        fieldName={field}
-                      />
-                    ))}
-                  </TableColumn>
-
-                  {/* Botones solo en la tabla de detalles */}
+                  {Object.entries(ROWS_CONFIG).map(([section, rows]) => (
+                    <TableColumn key={section}>
+                      <TableTitle>
+                        {section === "citaInfo"
+                          ? "Info Cita"
+                          : section === "pacienteInfo"
+                          ? "Info Paciente"
+                          : "Info Adicional"}
+                      </TableTitle>
+                      {rows.map(({ label, field }) => (
+                        <EditableRow
+                          key={field}
+                          label={label}
+                          value={
+                            field === "fecha"
+                              ? moment(appointment[field]).format("DD/MM/YYYY")
+                              : appointment[field]
+                          }
+                          isEditing={editingAppointmentId === appointment._id}
+                          onChange={handleInputChange}
+                          fieldName={field}
+                        />
+                      ))}
+                    </TableColumn>
+                  ))}
                   <ButtonContainer>
                     <Button onClick={() => handleDelete(appointment._id)}>
                       Eliminar cita
